#003 melhorias front

+++ resources/views/cliente/listar.blade.php
@@ -5,28 +5,37 @@
 @endsection
 
 @section('conteudo')
-    <table id="tabela_clientes" class="table table-striped table-bordered">
-        <thead class="thead-dark">
-            <tr>
-                <th>Nome</th>
-                <th>CPF/CNPJ</th>
-                <th>Editar</th>
-                <th>Deletar</th>
-            </tr>
-        </thead>
-        <tbody>
-        </tbody>
-        <tfoot>
-            <tr>
-                <td colspan="4"><a href="./pagina_cadastrar" class="btn btn-info">Adicionar Cliente</a></td>
-            </tr>
-        </tfoot>
-    </table>
+    <div class="table-responsive">
+        <table id="tabela_clientes" class="table table-striped table-bordered">
+            <thead class="thead-dark">
+                <tr class="text-center">
+                    <th>Nome</th>
+                    <th>CPF/CNPJ</th>
+                    <th>Editar</th>
+                    <th>Deletar</th>
+                </tr>
+            </thead>
+            <tbody>
+            </tbody>
+            <tfoot>
+                <tr>
+                    <td colspan="3">
+                        <a href="./cadastrar_cliente" class="btn btn-info">
+                            Adicionar Cliente
+                            <i class="fas fa-user-plus"></i>
+                        </a>
+                    </td>
+                </tr>
+            </tfoot>
+        </table>
+    </div>
 
     <script>
         var corpo_tabela = $("#tabela_clientes tbody");
         function carrega_dados()
         {
+            $("#gif_carregando").css('display', 'block');
+
             $.ajax({
                 url: './api/listar_cliente',
                 type: 'POST',
@@ -44,29 +53,26 @@
                     let coluna_cpf_cnpj = $("<td></td>");
                         coluna_cpf_cnpj.append(this.cpf_cnpj);
 
-                    let link_editar = './pagina_editar/' + this.id_cliente;
+                    let link_editar = './editar_cliente/' + this.id_cliente;
                     let btn_editar = $("<a></a>");
                         btn_editar.attr({
                             class: 'btn btn-warning',
                             href: link_editar
                         });
-                        btn_editar.append("Editar");
+                        btn_editar.html('Editar <i class="fas fa-edit"></i>');
 
                     let btn_excluir = $("<a></a>");
                         btn_excluir.attr('class', 'btn btn-danger');
-                        btn_excluir.append("Excluir");
+                        btn_excluir.html('Excluir <i class="fas fa-trash-alt"></i>');
                         btn_excluir.click(deleta_linha);
 
-                    let coluna_editar = $("<td></td>");
-                        coluna_editar.append(btn_editar);
-
-                    let coluna_deletar = $("<td></td>");
-                        coluna_deletar.append(btn_excluir);
+                    let coluna_acao = $("<td class='text-center'></td>");
+                        coluna_acao.append(btn_editar);
+                        coluna_acao.append(btn_excluir);
 
                     linha.append(coluna_nome);
                     linha.append(coluna_cpf_cnpj);
-                    linha.append(coluna_editar);
-                    linha.append(coluna_deletar);
+                    linha.append(coluna_acao);
 
                     corpo_tabela.append(linha);
 
@@ -74,7 +80,14 @@
                 
             })
             .fail(function(erro) {
-                console.log(erro);
+                swal({
+                    title: "Erro!",
+                    text: "Não foi possível carregar os clientes, verifique sua conexão e tente novamente!",
+                    icon: "error"
+                });
+            })
+            .always(function() {
+                $("#gif_carregando").css('display', 'none');
             });
             
         }
@@ -88,20 +101,54 @@
             let linha = $(this).parent().parent();
             let id_linha = linha.attr('id');
 
-            $.ajax({
-                url: './api/excluir_cliente',
-                type: 'DELETE',
-                dataType: 'json',
-                data: {id: id_linha, token: '{{$token}}'},
+            swal({
+                title: "Atenção!",
+                text: "A exclusão é permanente, deseja excluir registro?",
+                icon: "warning",
+                dangerMode: true,
+                buttons: {
+                    cancel: {
+                        text: "Cancelar",
+                        value: null,
+                        visible: true,
+                        className: ""
+                    },
+                    confirm: {
+                        text: "Excluir",
+                        value: true,
+                        visible: true,
+                        className: ""
+                    }
+                },
             })
-            .done(function() {
-                console.log("success");
-                linha.remove();
+            .then((escolha) => {
+                if (escolha) {
+                    
+                    $("#gif_carregando").css('display', 'block');
 
-            })
-            .fail(function(error) {
-                console.log(error);
-            })
+                    $.ajax({
+                        url: './api/excluir_cliente',
+                        type: 'DELETE',
+                        dataType: 'json',
+                        data: {id: id_linha, token: '{{$token}}'},
+                    })
+                    .done(function() {
+                        linha.remove();
+
+                    })
+                    .fail(function(error) {
+                        console.log(error);
+                        swal({
+                            title: "Erro",
+                            text: "Não foi possível excluir o registro!",
+                            icon: "error"
+                        });
+                    })
+                    .always(function() {
+                        $("#gif_carregando").css('display', 'none');
+                    });
+                }
+            });
         }
 
     </script>
