+++ resources/views/produto/listar.blade.php
@@ -5,29 +5,37 @@
 @endsection
 
 @section('conteudo')
-    <table id="tabela_clientes" class="table table-striped table-bordered">
-        <thead class="thead-dark">
-            <tr>
-                <th>Produto</th>
-                <th>Descrição</th>
-                <th>Valor (R$)</th>
-                <th>Editar</th>
-                <th>Deletar</th>
-            </tr>
-        </thead>
-        <tbody>
-        </tbody>
-        <tfoot>
-            <tr>
-                <td colspan="6"><a href="./pagina_cadastrar_produto" class="btn btn-info">Adicionar Produto</a></td>
-            </tr>
-        </tfoot>
-    </table>
-
+    <div class="table-responsive">
+        <table id="tabela_clientes" class="table table-striped table-bordered">
+            <thead class="thead-dark">
+                <tr class="text-center">
+                    <th>Produto</th>
+                    <th>Descrição</th>
+                    <th>Valor (R$)</th>
+                    <th>Ação</th>
+                </tr>
+            </thead>
+            <tbody>
+            </tbody>
+            <tfoot>
+                <tr>
+                    <td colspan="4">
+                        <a href="./cadastrar_produto" class="btn btn-info">
+                            Adicionar Produto
+                            <i class="fas fa-plus-circle"></i>
+                        </a>
+                    </td>
+                </tr>
+            </tfoot>
+        </table>
+    </div>
+    
     <script>
         var corpo_tabela = $("#tabela_clientes tbody");
         function carrega_dados()
         {
+            $("#gif_carregando").css('display', 'block');
+
             $.ajax({
                 url: './api/listar_produto',
                 type: 'POST',
@@ -35,7 +43,6 @@
                 data: {token: '{{$token}}'},
             })
             .done(function(dados) {
-                console.log(dados);
 
                 $.each(dados, function() {
                     let linha = $("<tr></tr>");
@@ -47,30 +54,28 @@
                     let coluna_valor = $("<td></td>");
                         coluna_valor.append(this.valor.toLocaleString('pt-br', {minimumFractionDigits: 2}));
 
-                    let link_editar = './pagina_editar_produto/' + this.id_produto;
+                    let link_editar = './editar_produto/' + this.id_produto;
                     let btn_editar = $("<a></a>");
                         btn_editar.attr({
-                            class: 'btn btn-warning',
+                            class: 'btn btn-warning mb-1 mr-2',
                             href: link_editar
                         });
-                        btn_editar.append("Editar");
+                        btn_editar.html('Editar <i class="fas fa-edit"></i>');
 
                     let btn_excluir = $("<a></a>");
-                        btn_excluir.attr('class', 'btn btn-danger');
-                        btn_excluir.append("Excluir");
+                        btn_excluir.attr('class', 'btn btn-danger mb-1 mr-2');
+                        btn_excluir.attr('href', '#');
+                        btn_excluir.html('Excluir <i class="fas fa-trash-alt"></i>');
                         btn_excluir.click(deleta_linha);
 
-                    let coluna_editar = $("<td></td>");
-                        coluna_editar.append(btn_editar);
-
-                    let coluna_deletar = $("<td></td>");
-                        coluna_deletar.append(btn_excluir);
+                    let coluna_acao = $("<td class='text-center'></td>");
+                        coluna_acao.append(btn_editar);
+                        coluna_acao.append(btn_excluir);
 
                     linha.append(coluna_produto);
                     linha.append(coluna_descricao);
                     linha.append(coluna_valor);
-                    linha.append(coluna_editar);
-                    linha.append(coluna_deletar);
+                    linha.append(coluna_acao);
 
                     corpo_tabela.append(linha);
 
@@ -78,7 +83,14 @@
                 
             })
             .fail(function(erro) {
-                console.log(erro);
+                swal({
+                    title: "Erro!",
+                    text: "Não foi possível carregar os produtos, verifique sua conexão e tente novamente!",
+                    icon: "error"
+                });
+            })
+            .always(function() {
+                $("#gif_carregando").css('display', 'none');
             });
             
         }
@@ -92,20 +104,54 @@
             let linha = $(this).parent().parent();
             let id_linha = linha.attr('id');
 
-            $.ajax({
-                url: './api/excluir_produto',
-                type: 'DELETE',
-                dataType: 'json',
-                data: {id: id_linha, token: '{{$token}}'},
+            swal({
+                title: "Atenção!",
+                text: "A exclusão é permanente, deseja excluir registro?",
+                icon: "warning",
+                dangerMode: true,
+                buttons: {
+                    cancel: {
+                        text: "Cancelar",
+                        value: null,
+                        visible: true,
+                        className: ""
+                    },
+                    confirm: {
+                        text: "Excluir",
+                        value: true,
+                        visible: true,
+                        className: ""
+                    }
+                },
             })
-            .done(function() {
-                console.log("success");
-                linha.remove();
+            .then((escolha) => {
+                if (escolha) {
+                    $("#gif_carregando").css('display', 'block');
+                    
+                    $.ajax({
+                        url: './api/excluir_produto',
+                        type: 'DELETE',
+                        dataType: 'json',
+                        data: {id: id_linha, token: '{{$token}}'},
+                    })
+                    .done(function() {
+                        linha.remove();
 
-            })
-            .fail(function(error) {
-                console.log(error);
-            })
+                    })
+                    .fail(function(error) {
+                        swal({
+                            title: "Erro",
+                            text: "Não foi possível excluir o registro!",
+                            icon: "error"
+                        });
+                    })
+                    .always(function() {
+                        $("#gif_carregando").css('display', 'none');
+                    });
+
+                }
+            });
+            
         }
 
     </script>
